ci: fix gcc conversion error
../src/handmadehero_asset.c: In function ‘LoadWav’:
../src/handmadehero_asset.c:513:24: error: conversion from ‘long unsigned int’ to ‘u32’ {aka ‘unsigned int’} may change value [-Werror=conversion]
  513 |   result.sampleCount = sampleDataSize / (fmt->numChannels * sizeof(u16));
      |                        ^~~~~~~~~~~~~~

+++ src/handmadehero_asset.c
@@ -510,7 +510,7 @@ LoadWav(pfnPlatformReadEntireFile PlatformReadEntireFile, char *filename)
   assert(fmt && sampleData);
 
   result.channelCount = fmt->numChannels;
-  result.sampleCount = sampleDataSize / (fmt->numChannels * sizeof(u16));
+  result.sampleCount = sampleDataSize / (u32)(fmt->numChannels * sizeof(u16));
   switch (fmt->numChannels) {
   case 1:
     result.samples[0] = sampleData;
